requirements.txt
appnope==0.1.3 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Darwin" or python_version >= "3.10" and python_version < "4.0" and sys_platform == "darwin"
asttokens==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
backcall==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
blinker==1.6.2 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0" and implementation_name == "pypy"
click==8.1.5 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows" or python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
comm==0.1.3 ; python_version >= "3.10" and python_version < "4.0"
debugpy==1.6.7 ; python_version >= "3.10" and python_version < "4.0"
decorator==5.1.1 ; python_version >= "3.10" and python_version < "4.0"
executing==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
flask-cors==4.0.0 ; python_version >= "3.10" and python_version < "4.0"
flask==2.3.2 ; python_version >= "3.10" and python_version < "4.0"
ipykernel==6.24.0 ; python_version >= "3.10" and python_version < "4.0"
ipython==8.14.0 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
jedi==0.18.2 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
jupyter-client==8.3.0 ; python_version >= "3.10" and python_version < "4.0"
jupyter-core==5.3.1 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "4.0"
matplotlib-inline==0.1.6 ; python_version >= "3.10" and python_version < "4.0"
nest-asyncio==1.5.6 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.10" and python_version < "4.0"
parso==0.8.3 ; python_version >= "3.10" and python_version < "4.0"
pexpect==4.8.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "win32"
pickleshare==0.7.5 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==3.9.1 ; python_version >= "3.10" and python_version < "4.0"
prompt-toolkit==3.0.39 ; python_version >= "3.10" and python_version < "4.0"
psutil==5.9.5 ; python_version >= "3.10" and python_version < "4.0"
ptyprocess==0.7.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "win32"
pure-eval==0.2.2 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0" and implementation_name == "pypy"
pygments==2.15.1 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
pywin32==306 ; sys_platform == "win32" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
pyzmq==25.1.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
snakemq==1.6 ; python_version >= "3.10" and python_version < "4.0"
stack-data==0.6.2 ; python_version >= "3.10" and python_version < "4.0"
tornado==6.3.2 ; python_version >= "3.10" and python_version < "4.0"
traitlets==5.9.0 ; python_version >= "3.10" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.10" and python_version < "4.0"
werkzeug==2.3.6 ; python_version >= "3.10" and python_version < "4.0"
